[REFACTOR] Move add challenge item flow into a new component
